requirements.txt
APScheduler==3.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
psycopg==3.0.16
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
vk-api==11.9.9
